app/requirements.txt
django
psycopg2-binary
django-environ
pillow
djangorestframework
drf-nested-routers